Refactored blockchain constructor, genesis, and creating new blocks

+++ src/main/java/edu/greenriver/sdev372/adamwinter/spring_project_372/models/IBlock.java
@@ -6,7 +6,7 @@ import java.util.Set;
 public interface IBlock {
 
     void setBlockId(long blockId);
-    int getBlockId();
+    long getBlockId();
     boolean addTransaction(ITransaction transaction);
     Set<ITransaction> getAllTransactions();
     String getHash() throws NoSuchAlgorithmException;
